requirements.txt
sounddevice>=0.5.0
pynput>=1.7.0
pyperclip>=1.8.0
pyautogui>=0.9.54
pystray>=0.19.0
Pillow>=10.0.0
numpy>=1.24.0
pyyaml>=6.0
scipy>=1.10.0
customtkinter>=5.2.0,<6.0.0
